requirements.txt
alembic==1.4.2
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
Bootstrap-Flask==1.4
certifi==2020.6.20
cffi==1.14.3
click==7.1.2
cryptography==3.1.1
dnspython==1.16.0
dominate==2.5.2
elasticsearch==7.9.1
email-validator==1.1.1
Flask==1.0.2
Flask-BabelEx==0.9.4
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.7
Flask-Logging==0.1.3
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.3
Flask-Uploads==0.2.1
Flask-User==1.0.2.2
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
migrate==0.3.8
passlib==1.7.2
pkg-resources==0.0.0
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2020.1
six==1.15.0
speaklater==1.3
SQLAlchemy==1.3.17
urllib3==1.25.10
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.1
